Criando o formulario de billingCycle e refatorando os Controllers genericos

+++ src/main/java/org/dougllas/mymoney/generic/AbstractCrudRestController.java
@@ -1,18 +1,18 @@
 package org.dougllas.mymoney.generic;
 
+import org.dougllas.mymoney.api.Response;
 import org.dougllas.mymoney.exceptionhandler.BadRequestException;
-import org.dougllas.mymoney.exceptionhandler.ResourceNotFountException;
 import org.dougllas.mymoney.exceptionhandler.ValidationHandlerController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.io.Serializable;
-import java.net.URI;
 
 public class AbstractCrudRestController<E extends Entity, ID extends Serializable, R extends JpaRepository> implements CrudController<E,ID>, ValidationHandlerController {
 
@@ -26,36 +26,42 @@ public class AbstractCrudRestController<E extends Entity, ID extends Serializabl
     @Override
     @GetMapping
     public ResponseEntity findAll() {
-        return new ResponseEntity(repository.findAll(), HttpStatus.OK);
+        return new ResponseEntity(Response.createResponse(repository.findAll()), HttpStatus.OK);
     }
 
     @Override
     @GetMapping("{id}")
-    public ResponseEntity findOne( @PathVariable("id") ID id) {
-        E entity = (E) repository.findOne(id);
-        if(entity == null){
-            throw new ResourceNotFountException("Item Não encontrado.");
-        }
-        return new ResponseEntity(entity, HttpStatus.OK);
+    public ResponseEntity findOne( @PathVariable("id") ID id, BindingResult result) {
+        E entity = validateResource(id, result);
+        Response<E> response = new Response();
+        if (handleErrors(result, response)) return ResponseEntity.badRequest().body(response);
+        return new ResponseEntity(Response.createResponse(entity), HttpStatus.OK);
     }
 
     @Override
     @PostMapping
-    public ResponseEntity save( @RequestBody @Valid E entity ) {
+    public ResponseEntity<Response<E>> save( @RequestBody @Valid E entity , BindingResult result ) {
+        Response<E> response = new Response();
+        if (handleErrors(result, response)) return ResponseEntity.badRequest().body(response);
         repository.save(entity);
-        URI uri = ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(entity.getId())
-                .toUri();
+        return ResponseEntity.ok(Response.createResponse(entity));
+    }
 
-        return ResponseEntity.created(uri).build();
+    public boolean handleErrors(BindingResult result, Response<E> response) {
+        if(result.hasErrors()){
+            result.getAllErrors().forEach(e -> response.getErrors().add(e.getDefaultMessage()));
+            return true;
+        }
+        return false;
     }
 
     @Override
     @PutMapping("{id}")
-    public ResponseEntity update( @RequestBody @Valid E entity, @PathVariable("id") ID id) {
-        validateResource(id);
+    public ResponseEntity update( @RequestBody @Valid E entity, @PathVariable("id") ID id, BindingResult bindingResult) {
+        validateResource(id,bindingResult);
+
+        Response<E> response = new Response();
+        if (handleErrors(bindingResult, response)) return ResponseEntity.badRequest().body(response);
 
         if(!entity.getId().equals(id)){
             throw new BadRequestException("Id passado está diferente do id da entidade.");
@@ -67,16 +73,19 @@ public class AbstractCrudRestController<E extends Entity, ID extends Serializabl
 
     @Override
     @DeleteMapping("{id}")
-    public ResponseEntity delete( @PathVariable("id") ID id) {
-        validateResource(id);
+    public ResponseEntity delete( @PathVariable("id") ID id, BindingResult bindingResult) {
+        validateResource(id,bindingResult);
+        Response<E> response = new Response();
+        if (handleErrors(bindingResult, response)) return ResponseEntity.badRequest().body(response);
         repository.delete(id);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
 
-    private void validateResource(ID id) {
+    private E validateResource(ID id, BindingResult bindingResult) {
         E existent = (E) repository.findOne(id);
         if(existent == null){
-            throw new ResourceNotFountException("Item Não encontrado.");
+            bindingResult.addError(new ObjectError("entidade","Item Não encontrado."));
         }
+        return existent;
     }
 }
